Added basic limit checking
to limit number of synced notes to no more than 250 per day. This is because Evernote has a limit of 250 emails per day for premium users. This is a quick fix, and better checking should be done, like asking the user if they are a free or premium user, and a default value for sync limit in the absence of user input.